Check if table exists before creating.

+++ src/database/migrations/2015_03_09_095623_create_coupon_user_table.php
@@ -12,14 +12,16 @@ class CreateCouponUserTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('coupon_user', function(Blueprint $table)
-		{
-			$table->increments('id');
-			$table->integer('coupon_id')->unsigned();
-            $table->integer('user_id')->unsigned();
-            $table->foreign('coupon_id')->references('id')->on('coupons')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-		});
+        if (! Schema::hasTable('coupon_user')) {
+            Schema::create('coupon_user', function(Blueprint $table)
+            {
+                $table->increments('id');
+                $table->integer('coupon_id')->unsigned();
+                $table->integer('user_id')->unsigned();
+                $table->foreign('coupon_id')->references('id')->on('coupons')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
 	}
 
 	/**
